requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
httpx>=0.27.0
pyyaml>=6.0
